Edits and corrections to threads section of the book
Antonio Diaz Diaz informed me that the statement "The number of threads
available to your operating system is usually double the number of physical
(hardware) cores in your CPU" only applies to some CPUs (for example
Intel's CPUs that have Hyper-threading). So this statment was corrected and
more generally, I went through the whole "Threads in Gnuastro" section and
made some edits, corrections and additions. In particular on describing the
unique features of Make for scientific research.