Segfault in mesh interpolation corrected
When the input image is very small (about the same size as the mesh grid),
then there will only be one mesh. Until now, the neighboring meshes would
be checked in any case, so when there is only one mesh (no neighbors),
there was a segmentation fault. With this commit, the neighboring meshes
are only checked when `numngb' is not zero.

This bug was reported by Nicolas Bouché, so the `THANKS' file was updated.

This fixes bug #49588.