Four common options now usable in config files
Not all the common options could be used (read from or written to) in
configuration files until now. So with this commit, the four common options
`--quiet', `--onlydirconf', `--dontdelete', and `--keepinputdir' are now
read from and written to configuration files.

Note that options like `--cite', `--printparams', `--setusrconf',
`--setdirconf', along with `--help', `--usage' and `--version' don't need
to be put in the configuration files, so not all common options needed to
be included there.

This finishes task #14219.