Compression suffixes accounted for in automatic output
The `gal_checkset_automatic_output' function is in charge of setting the
filename of the output filename when the user hasn't specified any.

It starts from the end of the input file name and replaces everything
including and after the first `.' it encounters. This works on ordinary
`.fits' files, but not on inputs like `.fits.gz' (that CFITSIO also
accepts). So for example ConvertType's default output (to convert to Jpeg)
on such a file would end with `.fits.jpg'. With this commit, some further
checks are added to see if the text after the last `.' is a compression
suffix or not. If so, it continues going down the filename in search of the
next `.'. This is not the most elegent solution, but with the current
implementation of `gal_checkset_automatic_output', this was the best
solution I could find right now.

This fixes bug #49629.