Configure date and time no longer kept or written
From Debian's reproducible builds reports, apparently the only
reproducibility problem in Gnuastro is the configure date and time that is
recorded at configure time and passed to all the executables.

With all other things being equal (same tar-ball, same dependencies and
etc), the configuration time is completely arbitary/irrelevant and since
Gnuastro was first started, I haven't found myself actually paying any
attention to it or wanting to use it. So I feel it is good to remove it
with the benefit of having a reproducible build.