Renamed lib/gnuastro.h to lib/config.h
The header file containing the Gnuastro installation configurations was
called with `gnuastro/gnuastro.h', or its macros were called
`GAL_GNUASTRO_*', this isn't too clear and could be confusing. Since
practically the contents of this header are the same as those in `config.h'
when building Gnuastro, the header name was changed to
`gnuastro/config.h'. The name is clear and unambiguous both when including
and also when calling the macros (or possibly functions later).