Defined new gnuastro/data.h header
The new `lib/gnuastro/data.h' header is now defined with a generic (and
currently untested) C structure. In the next commits it will be slowly
implemented in all the libraries and utilities. In this commit, the blank
values for the different datatypes have been moved from the `fits.h' header
to `data.h' and they are corrected within Gnuastro.