Corrected bad use of type in MakeProfiles
In MakeProfiles, the `type' name was used in two contexts: one for the
datatype (which is expected), but another case was in the radial function
type! This could cause confusion and was a bad practice. So with this
commit, all occurances of the latter case have been changed to `func'.