Arithmetic with internal type conversion
Arithmetic now uses the internal type conversion of C on arrays on internal
binary operators when the types are different.  Until now, it would convert
them to the same type and then do the operation. But now, there is no need
to do any conversion. This will speed up the process, but it is a strong
burden on the compilation, so the compilation can become very slow,
especially, when all the types are desired, on my computer it took about 40
minutes to compile `data.o' and `data.c'. So some configure time options
have been added to allow the users to choose the non-basic types if they
want them.

Don't forget to add/describe the configure time options in the final
manual, and also the new macros in `gnuastro/config.h' describing the
different arithmetic types.