Operators treated as macros, not string
The different operators are now recorgnized as macros in
`gal_data_arithmetic'. Until now, they were treated as strings which made
it a little slow and hard to generalize.

Also the `!=' operator is now called `ne' in Arithmetic, until now it was
called `neq'. This follows the names of the comparison operators in the
shell.