Corrected bad status usage in wcsp2s and wcss2p
The status argument to WCSLIB's `wcsp2s' and `wcss2p' functions is not a
simple variable, but an array (WCSLIB apparently writes a status value for
each coordinate). So on some compilers, this caused a segmentation error
when more than one coordinate was input.

This bug was reported by Christopher Willmer, he also patiently helped in
finding its cause and confirming the solution.

This fixes bug #49752.