Data structure mmap information updated
As correctly suggested by Antonio Diaz Diaz, the `mmapped' variable in the
data structure is redundant and will only cause bugs. To check if the array
is allocated or mapped, it is sufficient to check the `mmapname' pointer:
if its `NULL', then the array is allocated. So this variable was removed
from the structure.

On the other hand, in many cases, data structures will be created/derived
from existing data structures, and it is a burden to pass around the the
`minmapsize' number. So this variable is now defined in the datastructure
and derivative data structures will be able to specify if they should be
allocated or mmap'd.

Some minor corrections were also made in the manual (in the installation
directory).