Binary operator arithmetic works on uncompiled types
Until now, if the inputs to the binary operators were an uncompiled type,
`gal_data_arithmetic' would stop with an error. But now, there are checks
before the actual operation and if the types of any of the inputs or output
are not compiled, they will be convereted to an existing larger datatype.

In `configure.ac', the check to set the printed value in the configure
message was moved immediately after the value is set, so nothing in the
middle can possibly change it, making the printed value more trustable.

The new function `gal_data_type_string' was also created so it will be easy
to print readable debugging messages, but it can be used in other contexts
too.