Use of function instead of macros for binary operators
Until now, it was planned for all the operators to be called as
pre-processor macros from the top level `gal_data_arithmetic'
function. This made the code slower to compile and was also hard to manage
since a lot of the process had to be in macros.

With this commit, we have a separate function for ordinary (non-bitwise)
functions which is in its own `.c' file and thus compiled independently
from the functions of other operators. Also since the use of macros is
drastically decreased, the compile time is much more reasonable.

The idea for this change came during the preparations for the
implementation of bitwise operators, they are also a significant number and
having them all be compiled from one function along with the rest of the
operators would be really slow.