Corrected occurance of "remainder" with modulo
The main name for the `%' operator is "modulo" (according to Wikipedia).
However, in my first implementation, I used the term "remainder" in the
macros and book. Later, it was corrected, but apparently one occurance in a
macro and two occurances in the book slipped through my fingers. They are
now corrected.