Type conversion operators in Arithmetic
In this branch, we are working on a root data structure to facilitate the
manipulation of input, output and generally internal data processing. In
particular, a data array pointer is generally stored as a `void *' and
there is an element in the structure to keep its type. This will enable a
high-level usage of the data structure independent of the actual type of
the data.

Until now, Arithmetic would convert all input numbers and arrays into
double precision floating point numbers and do the processing on them. But
as the datasets get larger, this process (both the conversion and the
processing) can make things slow when the input data aren't `double' (most
astronomical images are stored in single precision floating point types).

Thanks to the new data type, it is now possible to actually do the
arithmetic on arrays and numbers with their original input types. Since the
original types are preserved in Arithmetic now, in some cases, it might be
necessary to explicitly change the input types to another type (for example
division of two integer arrays). With this commit the necessary type
operators (based on the CFITSIO image types) have been defined and are now
usable. The explanation of the operators is now also within the manual.

Other minor corrections:

 - `data_arithmetic_not' didn't free the input if the caller asked for
   it. So this is also now corrected.

 - The part of `gal_data_alloc' which copied the WCS structure is now taken
   out as a separate `gal_data_copy_wcs', to generally copy the WCS of any
   data structure into another.

This finishes task #14254.