Removed `anyblank' from datastructure, several new operators
The `data-arithmetic-unary.[ch]' library files were changed to
`data-arithmetic-other.[ch]', since it will host functions for a more wider
class of operators. This was mainly prompted by the new
`data_arithmetic_unary_function_f' and `data-arithmetic_binary_function_f'
functions.

The `gal_data_string_to_number' can now be told to only read the number as
a floating point by adding an `f' to the end of the string.

`gal_data_operator_string' is a new function to return the string
equivalent of an operator, mainly useful for error messages.