New Data types section in book, bitwise not added
The explanation that was previously under the `uchar' operator in
Arithmetic on data types has been moved to its own separate subsection, and
made more clear.

Also, the bitwise not (or one's component) operator is now implemented in
`gal_data_arithmetic' and thus the Arithmetic program.

Also, in preparation for more detailed use of the root data structure, a
new `next' pointer is now added to it so the datatypes can be used as
linked lists.