Absolute value operator in gal_data_arithmetic
The absolute operator is now implemented in `gal_data_arithmetic' and thus
the Arithmetic program. Since there are multiple functions to get the
absolute value in the C library based on the type, each data structure type
uses the best available function.

Also, the `operator' argument was irrelevant in `where', it has been
removed.