Minimum and maximum value operators implemented
The minimum and maximum value operators are now available to the
`gal_data_arithmetic' library function and thus to Arithmetic. As part of
implementing these functions, the old `data_arithmetic_unary_function_f'
was removed and is now generally usable for any type. Depending on the
operator, control will either be taken to functions to work on each element
(for example taking the log of every pixel), or to work on the full array
(like minimum or maximum values).

Other smaller work that is done with this commit:

 - When the output of Arithmetic is a single element data structure, it is
   printed as a number, not saved as a FITS file (like before this branch
   started).

 - The unused `operator' argument to `data_arithmetic_bitwise_not' was
   removed.

 - The two new `gal_data_type_min' and `gal_data_type_max' functions now
   return the maximum value of the given type in a `void *' pointer.