All old arithmetic operators are now implemented
The final remaining operators in Arithmetic were those that worked on any
number of operands (like `min', or `median'). In the old versions, these
operators would pop every operand. But now, they read one operand first
(which must be a number and will tell the operator how many operands to
pop). Then it will pop the given number of operands and define a linked
list of data structures which will be fed into `gal_data_arithmetic' for
the actual operation.

The Arithmetic chapter of the book was also updated to include all the
updates. With this commit, the first draft of Arithmetic using the new root
data structure has been implemented and we can go onto the rest of the
programs/librarires.