More efficient macro implementation for binary operators
Previsouly, the macro implementation for the various types of binary
operators also iterated (with switch expressions) over types for the
output. However, the output type is determined from the input types or the
operators, it isn't random. With this commit, this fact is taken into
account in defining the macros of `data-arithmetic_binary.c' and
`data_arithmetic_onlyint.c'. With the full set of types enabled,
compilation time on my system decreased from about 40 minutes to 17
minutes.

The book was also modified/updated/edited to better explain the issues.