Data structure with name, units, comments and status
In preparation for reading and writing FITS tables (and also more
customizable image keyword outputs), the data structure now has several new
elements: `status', `name', `units', and `comment'. The last three are
strings.

With these new elements the old `gal_fits_key' structure was no longer
necessary for the `gal_fits_read_keywords' function. It can now directly
save its results to the root data structure instead of requiring a unique
structure for itself. `gal_fits_read_keywords' is now also able to read the
units and comments of a keyword, which can be very useful.

If the `BUNIT' and `EXTNAME' keywords are present in a FITS file,
`gal_fits_read_img_hdu' will also read them store them in the structure for
use by the programs.

With the addition of these keywords, and also starting work on string
datasets, the `gal_data_free' function was also updated to free all the
arrays pointed to by the array of strings. Also, to allow arrays of data
structures, freeing the actual data structure is now optional with this
function: a new argument is added to it and corrected in all the programs
that use it so far.

With the same line of reasoning as above (to allow working on arrays of
data structures), the lower level parts of `gal_data_alloc' have moved to
the new `gal_data_initialize' function which will assume an already
allocated data structure and just initialize its values.

The old `txtarray.[ch]' files have been deleted and replaced with a new
`txt.[ch]' for input and output on text files (to be similar to `fits.h').