Starting to use the new data structure to Table
Reading and writing columns of data are a very important part of most
programs and also for the libraries. So Table was chosen as the second
program to migrate to the new data structure. Each column can be read as
one data structure and all the necessary information is stored in it. A new
`table.h' library header was defined and practically all of what was in the
Table program is now slowly shifting in there so all the programs and also
library users can benefit from easy reading and writing of columns.

For now, Table will print the full information of a given FITS table. It
can also select the columns based on the `--column' option as before (but
much better). Similar the AWK, if the users want regular expression search,
they should put the value in to slashes like `/ /', and all of AWK's regexp
syntax is applied. It is also possible to choose which information of a
column should be checked: the name, the units or the comments. While the
columns can be chosen, work is still on-going on the library to finalize
the reading and writing of table columns.

To help in re-writing Table, its folder was essentially copied outside and
all its functions cleaned, so we can cleanly re-write all those important
functions into the library with the new data structure.