Any number of searched columns from FITS are read
The high level `gal_table_read_cols' function can now read any number of
columns from a given FITS table. Also the display format FITS header
keywords are now read into the data structure when they are present. The
great advantage of this is that since it is now implemented as a library,
it can be used by any of the programs (and even users of the libraries) to
greatly simplify the process of reading the input data into the program for
further analysis.