Corrected incrementation issue with and and or operators
The `&&' and `||' operators will not necessarily read their second
operand. So the default `BINARY_OP_OT_RT_LT_SET' wouldn't read the second
array sometimes and this caused unsynchronized incrementations. So with
this commit we have a new `BINARY_OP_INCR_OT_RT_LT_SET' macro specially for
these types of operators. In this new macro, the second operand is
incremented in a a separate expression.

Also, the freeing of the keys linked list after reading keywords in
`gal_fits_read_img_hdu' was moved after `gal_data_alloc' so the values are
freed. In the old way, the two strings were left unfreed until the end of
the program, because they were artificially set to NULL (mistakenly
assuming that we won't need to copy them, while `gal_data_alloc' copies
them).