Minimum sharp PSF spectrum is now an option not macro
The minimum sharp PSF's spectrum value to use in Convolve's deconvolution
was a macro until now, making it very hard to play with. So now Convolve
accepts this value as an option. For the job, a new
`gal_checkset_double_l_0_s_1' was defined in the internal Gnuastro library
to directly read a double type value and check if it is between 0 and 1.