Small correction to section crop in ImageCrop
In the `Crop section syntax' of the manual, it was stated that "there are
no -sub-pixels in ImageCrop". This can cause confusion, because in defining
the pixel centers to crop from (not in section crop mode), we do allow
sub-pixel positioning. In general, it is only the cropping that is not done
in sub-pixels, not the positions). For the `--section' option, this is
unique (and it only accepts integers). The corrected paragraph should
(hopefully) not cause this confusion.