Missing --numbered option in git shortlog for generating AUTHORS.
The source package script `genauthors' uses the command
`git shortlog' and generates the file AUTHORS, claiming that
the authors are ordered by their number of commits. But the
ordering option was forgotten :). TODO: There is no weighting
option to differentiate minor commits from significant
commits. A weighting by the number of lines changed would
partly overcome this, but the significance of edits is not
always proportional to the number of lines edited!