Arithmetic functions/macros out of data.h
The arithmetic operators were declared in `gnuastro/data.h'. However, this
header is only for low-level operations on (a single) data structure. Its
thus not good to complicate this header/library with arithmetic operations
on (possibly) multiple data types. Thus with this commit, the
arithmetic-related functions and macros in `gnuastro/arithmetic.h' are now
in a more specific `gnuastro/arithmetic.h' header.