Column info format for writing ASCII tables
The old ASCII column information was very rudimentary and included the
exact `printf' formatting for that column. However, for all numerical types
this detailed information is irrelevant and for strings, we only need the
width. So now, the column information follows this format (where all the
values written in capital letters can be randomly set).

  # Column NUM: NAME  [UNIT, TYPE] COMMENT

`NAME', `UNIT', and `COMMENT' are just strings and can have any ASCII
characters except the new-line and `[' for NAME, and `,' for UNIT (which
specify the limits of the respective field in the last two cases). Also,
any number of spaces between each element and the next separator is
acceptable.

During the process the following two issues were also addressed:

 * A notice was put in `gal_fits_table_write' to inform the user that this
   function is not ready/implemented yet.

 * Report an error if the user asks for column number 0. The column
   numbering starts from one in Gnuastro. Until now, the numberical value
   was stored in the same `if' statement, so when the input value was `0',
   it would fail and go onto making an exact match check. Now, the long
   type value is read before the check, so if it is zero, control goes into
   the numerical/integer branch.

 * In making an exact match, until now there was no check, so if there was
   no exact match, it would just ignore it. But now if there is at least
   not one match, there will be an error.

 * In `configure.ac', the result for `HAVE_BIN_OP_LONGLONG', was mistakenly
   written as `HAVE_BIN_OP_LONG', so in the output `lib/config.h', it would
   be blank. This is now corrected.