Column info read from comments in ASCII tables
The column information can now be read from a properly formatted comment in
an ASCII txt table file as described in the previous commit. It is the same
format that `lib/txt.c' will write a table to an ASCII table. This makes it
easy to read and write tables to and from FITS formats while preseving
information like column name, units, and etc.

Some other things that were done in this commit:

 - To make the type reading and writing easier, a new
   `gal_data_string_as_type' was defined in `data.c' to read a string and
   specify the type it refers to. The old `gal_data_type_string' was also
   renamed to `gal_data_type_as_string' since it does the opposite.

 - The `gal_data_copy_to_new_type' is now in `data.c' and the old
   `data-copy.[ch]' files were removed.