ASCII table information fully ready for selection
The column information of an ASCII table is now fully read into an array of
data structures in the internal format making further selection of columns
very easy. It is exactly the same output as that from a FITS table, so all
programs that need column inputs can now read from an ASCII table or a FITS
table.

In the process the following changes were also made:

 - When the caller wants a cleared memory and the memory has to be mmap'd,
   it wouldn't get cleared, it was just allocated. `gal_data_mmap' now
   clears the memory if the user asked for it.

 - Until now, `gal_data_initialize' implicitly assumed an `ndim>0'. But now
   it can also deal with an `ndim==0', for when there is no data to be put
   into the structure.