The table library can read ASCII inputs
`lib/txt.c' can now fully read the desired columns in an ASCII table (with
whitespace characters or comma as delimiters). It will also check for blank
values (specified in the information comments) and store them as Gnuastro's
internal blank values.

When reading the number of columns in a table, previously we would put that
in the `size' element of each column's data structure. But that greatly
complicates things. Because now, in the ascii tables, we store the user's
given blank value in the array of the data structure. So having `size' be
interpretted as something else, would mix different contexts. Thus, the
`gal_fits_table_info' and `gal_fits_table_info' functions now return the
number of rows as a single number, and leave the `size' element of the
structure untouched. The respective table reading functions also accept
this number as input.

While the inputs are now complete for a first draft, the printing of
outputs still doesn't treat blank values properly.

Other minor corrections:

 - Table's `--help' output was updated, with more clarifications.

 - When printing information, Table will also print the number of rows.

 - In freeing an array of strings, previously we were incrementing
   pointers. But when we have NULL pointers (for a blank string), the
   incrementing will cause trouble. So now we use indices to free the array
   of pointers.

 - As a test convention, counters that start from one will be called
   `number' and the `n' variable will be used for them. And, counters that
   start from zero will be called `index' and the `i' variable will be used
   for them.