Table library prints ASCII columns with blanks
Until now, the information for columns with blank values weren't correctly
written in an ASCII table information comment. But now, if a column
contains blank values, the value will be printed in the information column,
making it easy for parsers to specify which fields are blank. Also, when
dealing with blank strings, we now print a `blank' placeholder.

During the process, a problem in `gal_data_has_blank' was corrected: the
value to the ending pointers was based on `data->array' (which has `void *'
type), so adding `data->size' to it was useless, the ending pointers are
now defined based on the starting pointers (which have a type and thus a
size) to use.