Corrections to FITS table reading
During the FITS table reading, it is much more convenient to have an array
of data structures (with column information) to easily get to the desired
column information. But there was no systematic way to allocate an array of
data structures, so a new `gal_data_calloc_dataarray' was defined to make
the job of allocating (and initializing the values) easy.

Other minor corrections:

 - The array keeping the column information is now uniformly named
   `allcols' in all `lib/table.c', `lib/fits.c', and `lib/txt.c'.

 - `gal_fits_table_read' would pop elements from `indexll'. But it is
   better for reading/debugging that memory be freed in the same function
   they were allocated. So this function now only goes over the list, it
   doesn't pop/free the allocations. This was applied in `lib/txt.c', so
   when reading a FITS table, we would get segmentation errors.