Table's output file type set common sense
Previously, when the (old) `--outtabletype' option was given, the output
was forced to be a file (automatic output would be triggered when no output
filename was given). But this is not too natural, because you might have a
special FITS type as a preference, and thus put it in your configuration
file. But then, you could never pipe your outputs to other programs (show
the table on stdout).

So now, we have a `--tabletype' option, with the same acceptable values as
before, but now it won't be used to set a filename. In short, Table won't
be using automatic output: when no output type is given, the table will be
printed in stdout. The value to `--tabletype' is only relevant for table
formats that can't directly be determined from the filename.

This option will later be moved as a common option, for use by all the
programs that print tables as output. As a preparation, the function that
determines the recognized internal macro for the table type from a
string. The same will happen for the `searchin' parameter, so the old
`gal_table_searchin_from_str' was also renamed to
`gal_table_string_to_searchin' to fit the same naming style as the new
`gal_table_string_to_type'.

Also, `strtol' saves values as zero when they couldn't be read. This would
cause problems when setting widths. So now, in plain text printing, when
the width is smaller and equal to zero, the default widths will be evoked,
not when it is just smaller than zero.