With no columns, Table program will print all columns
When no columns are selected for Gnuastro's Table program, it will get the
table information, artifically make a list of all the columns, so the rest
of the program can be the same as before.

This feature wasn't implemented in the library because it is not a generic
thing: other programs will also be using the library and when there is no
columns, it is not acceptable to choose everything.