Problem in reading blank FITS ASCII table fixed
With the great help and guidance of William Pence (autor of CFITSIO) the
problems in FITS ASCII tables discussed in commit 232e059 (First draft of
FITS table writing function complete) were corrected. The main issues can
be summarized as bellow:

 - In the ASCII format, we need the `TNULLn' keyword for all types as a
   string. While in the binary format, we just need the `TNULLn' keywords
   for the integer and string types. Previously I had confused the two and
   didn't put the `TNULLn' keyword for floating types in ASCII tables.

 - All columns that need a `TNULLn' (integer and string types) must have it
   as long as at least one column has a blank value. So for the time being
   we are simply including the `TNULLn's for all integer and string type
   columns, not just those that have a blank value.

William Pence pointed to an important point: that `nan' values cannot be
robustly readed on many machines/compilers, so he suggested to use numbers
that are greatly out of range instead of a `NaN' value. This is an
important issue and must be addressed, but I currently don't have time. So
for the time being, Gnuastro also imports Gnulib's `strtod' module which
accounts for many of those failures in other systems. So atleast as long as
people use Gnuastro on those systems, they won't have any problem.

For the time being, if users need the ASCII tables (FITS or plain text) for
programs on systems that can't read NaN values properly, they can simply
use a `sed' command or a search and replace to change all the `nan' values
into some absurd numerical value. This will also change the plain text
comment, or FITS ASCII table `TNULLn' keyword, so the table is still
exactly readable in Gnuastro.

Work has also started on a generic table (which contains blank columns,
string columns and missing information), but it is not yet finished.

Some further modifications during this commit:

 - When printing table information, we use `GAL_DATA_BLANK_STRING', not the
   `N/A' string.

 - Since `strtod' is much more robust than `strtof' (and is supported in
   Gnulib), all invokations of `strtof' have been replaced with `strtod'.

 - `gal_data_blank_as_string' now also gets a `width' argument to write the
   blank value in. This is currently necessary for the FITS ASCII tables as
   discussed above.

 - When CFITSIO doesn't recognize certain table column types, an error will
   be printed. Since Gnuastro's table reading/writing is more general than
   FITS, we need those types and can't simply remove them.

 - `gal_table_col_print_info' now keeps the width and precision values it
   finds into the actual data structure for easy use and access at later
   times.