FITS ascii and binary table I/O done with tests
The FITS standard doesn't have any code to store `signed char', `unsigned
short', or `unsigned int' types. So they suggest using special values for
the `TSCALn' (to scale a column values) and `TZEROn' (to shift the column
values) keywords to let CFITSIO know that we are dealing with one of these
types instead of their unsigned/signed counterparts.

While CFITSIO does do this internally, we also need to do this outside
CFITSIO because the space/type to keep the values must already be
allocated/known. So `gal_fits_table_info' now reads the `TZEROn' and
`TSCALn' keywords also and the new function
`fits_correct_bin_table_int_types' is in charge of checking these two
values and correcting the type if it is necessary. With this correction
(and those in the last commit), the example table in
`tests/table/table.txt' which contains all types and blank values and etc
can now be read as an ASCII or binary FITS file and converted back into a
plain text file.

The test table was also corrected to be usable with the default values as
discussed in the comments above the table. Also, two tests were added to
first read this table as an ASCII FITS table, then convert that back to a
plain text table (two similar tests with Binary FITS tables alreay
existed). All four of these tests now pass (on my system at least).