Stricter checking for polygon point list argument of imgcrop
The main idea is that every point is the floating number separated with ';'
or ',' or space from the next number. So, after converting a string to a
number the pointer should point to one of separators.  Otherwise the
floating point has some invalid characters.

Summary of rules of a good number (after conversion):

  1. Pointer can stop at the end of line (check for '\0').

  2. Pointer can stop at whitespace (check for isspace).

  3. Pointer can stop at a separator (check for ':' and ',').

Why check for space is considered to be OK. A user can set argument list in
two ways:

  1. Without any space, like "polygon=1.1,0.1:1.2,2.0".

  2. And using spaces to make it more readable, like:
     "polygon='1.0,1.5 : 1.5,2.6'".

This fixes bug #48978.