Table info printing in libraries, updates to Table program
Printing the information of a given table on the command line in a nicely
formatted manner is now defined as the library function
`gal_table_print_info'. Previously, it was only usable within the Table
program. This was done because it is a very convenient feature and will
certainly be useful for checks or debugging in many program or library
development scenarios.

As part of this process, the width of the columns when printing column
information is not a hard-coded value any more, `gal_table_print_info' will
check the values before hand and set the width based on the values.

Another major change in this commit is that when the output of
`gal_txt_table_write' is not a file (to be sent to `stdout'), the column
information and comments will not be printed. This was done to allow easy
piping to other programs like AWK, or sort. The book was also corrected to
fully explain this change in the Table program.