New elements for argp_option for new option management
This is the first commit for task #14300 ("Option management using new root
data structure."). The main purpose was to make it simple to read the
options and their values from the command-line and configuration files. The
way the system works now is to use `argp' for command-line options, and 3
functions in each program's `ui.c' to read the command-line. These three
functions are very repetative, and we have to keep the string value of each
option twice. This is hard to maintain and buggy.

To do this, in `bootstrap.conf', we add the two elements `value' and `type'
to the definition of the argp_option' in the bootstrapped `argp.h'. Through
`config.h', Gnulib gives its own headers higher precedence, so we can
ignore the (possibly existing) `argp.h' in the user's own system and use
these new elements. Also, the definition of `argp_option' is no longer
static, so we can pass its pointer to the libraries.

From the new data structure, we just needed the type macros, not a full
`gal_data_t'. Also, a new type was added for string linked lists, mainly
for this task (reading options), but later it can be useful to generally
define linked list types too and unify all the linked list functions in
`linkedlist.h' into one.