Single function to read all configuration files
This commit mostly completes task #14300 ("Option management using new root
data structure.", as with the last few commits). But it is not yet
complete, Jobs that can be done immediately after reading the configuration
(like printing values or citation and so on) aren't yet included, the
values have to be written into the program's main paramters structure and
also, the documentation needs to be updated.

Until now, reading the configuration files heavily relied on macros, but
now, there is a simple function that will do the job, so is much more
easier to read, understand and debug.

As part of this work, the following changes were also made:

 - `gal_linkedlist_add_to_stll' now has one more argument to see if the
   string should be allocated and copied, or should only the pointer be
   passed.

 - The final installation directory is now only necessary in the library
   installation, and not the whole of Gnuastro. So the `-DSYSCONFIG_DIR'
   `CPPFLAGS' value is now defined in the `AM_CPPFLAGS' of
   `lib/Makefile.am' and the programs can easily be compiled without it.

 - The `CURDIRCONFIG_DIR' is no longer included in `config.h' because it is
   easily derivable from the `PACKAGE' macro.

 - A new option `--config' was added so the users can specify any number of
   custom configuration files that will be read before the default ones.

 - The `onlydirconf' option was changed to `lastconfig'. The previous
   version was too specific. Now, which ever configuration file contains
   this option (even on the command-line) will be the last configuration
   file that is read.

 - The `gal_txt_line_stat' function and `gal_txt_line_status_enums' have
   now been moved into `gnuastro/txt.h' since they were also useful in
   `lib/options.c'.

 - The `gal_linkedlist_reverse_*' functions now check if the value is not
   NULL and also that there is more than one element in the list.