Option values now usable in program
The option values can now be used in the programs. There is still some
small amount of work to complete task #14300 ("Option management using new
root data structure."), but fortunately most of it is done. The only
remaining parts are writing the printed paramters to configuration files
when the `set*conf' options are given. Also, some polishing of the normal
parameter printing function (print comments, and the extra information).

The book was also updated to fit some of the parts, but one full reading
over the text has not been done yet to make it nice and homogenous.

To do this, the following changes were also made:

 - Most of the `uiparams' elements have been removed.

 - The option keys are now enumerator types. This allows easy checking of
   option array elements, instead of having to use `strcmp' on the names.

 - The citation printing is also now complete: each program has its own
   `bibtex' variable which should keep the BibTeX record(s).