Single correction of string keyword values in FITS tables
Unfortunately the CFITSIO functions that return the value of a string
keyword do so with the single quotation marks before and after the
string. This can cause problems when we need to read the string as in
`gal_fits_table_info'. Until now, for each keyword, this correction was
made separately, but now, when the first character of the keyword value is
a single quote('), then the string will be corrected, so no matter what the
option is, a unified (trust-worthy) pointer (`val') can be used.

Another important issue that was found by running `make check' in the
previous commit was a `double free' crash. The cause was that when
`gal_data_string_to_type' couldn't read a value, it would free the
allocated array, but it wouldn't reset it to NULL. So later steps would
also try to free it. This is directly related to the paragraph above:
because of the single quotation marks, `TNULLn' keywords in FITS ASCII
tables couldn't be read and because of this issue, the check would fail.

During this process the following corrections were also made:

 - The `ui_read_check_options' functions in each program's `ui.c' will be
   in charge of putting the options into the main program's data structure
   and also doing the sanity checks. So the old `sanitycheck' function in
   `ui.c' is no longer needed.

 - The old `ui_preparations' option will be called `ui_preparations' from
   now on.